web: try to display the RSS feed entries on one line each.
Change the date from "Sunday, September 2013" to "07 September"; it's a
shorter, yet correct, date format that carries everything one expects.
Keeping the year is of little use since it's very uncommon to have entries
from the previous year: it's not ambiguous.

The date was also always on a separate line; make that non-forced.

git-svn-id: https://svn.code.sf.net/p/mingw-w64/code@6241 4407c894-4637-0410-b4f5-ada5f102cad1